+++ assets/js/App/Main/Content/About/Description/Description.jsx
