+++ assets/js/App/Main/Content/About/Resume/wrapAsTab/wrapAsTab.jsx
