+++ assets/js/App/Main/Content/Blog/Blog.jsx
