+++ assets/js/App/Main/Content/Blog/Index/Index.jsx
