+++ assets/js/App/Main/Content/Blog/Index/PostCard/PostCard.jsx
