+++ assets/js/App/Main/Content/Blog/Index/PostCard/PostCard.less
@@ -1,0 +1,32 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  border-radius: 5px;
+  border: 1px solid #dadada;
+  margin-bottom: 10px;
+  padding: 10px;
+}
+.header {
+  font: 25px sans-serif;
+  padding: 0 25px;
+  font-family: Sniglet;
+  text-transform: uppercase;
+}
+.preview {
+  font-size: 2;
+}
+.tags {
+  display: none;
+}
+.show_more_button {
+  background-color: #eae7e7; /* Green */
+  border: none;
+  color: black;
+  padding: 5px 10px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 14px;
+  border-radius: 5px;
+  width: 100px;
+}
