+++ assets/js/App/Main/Content/Blog/Post/Post.jsx
@@ -1,0 +1,15 @@
+import React from "react"
+import { Markup } from 'interweave';
+
+import style from './Post.less'
+
+export default function Post({post}) {
+  return(
+    <div className={style.main}>
+      <div className={style.header}>
+        {post.header}
+      </div>
+      <Markup className={style.preview} content={post.mainPart} />
+    </div>
+  );
+}
