+++ assets/js/App/Main/Content/Content.jsx
@@ -5,21 +5,22 @@ import {
   Switch
 } from 'react-router-dom'
 
-import Blog from 'Components/Blog/Blog'
-import About from 'Components/About/About'
-import NewPost from 'Components/Blog/Post/NewPost/NewPost'
-import EditPost from 'Components/Blog/Post/EditPost/EditPost'
+import Blog     from './Blog/Blog'
+import About    from './About/About'
+import ShowPost from './Blog/Post/Show'
+import NewPost  from './Blog/Post/New'
+import EditPost from './Blog/Post/Edit'
 
 export default class Content extends React.Component {
   render() {
     const aboutMe = this.props.aboutMe;
-
     return(
       <Switch>
         <Route exact path="/" render={ () => <About aboutMe={aboutMe}/> }/>
-        <Route path="/blog" component={Blog}/>
-        <Route path="/posts/new" component={NewPost}/>
+        <Route path="/blog"           component={Blog}/>
+        <Route path="/posts/new"      component={NewPost}/>
         <Route path="/posts/:id/edit" component={EditPost}/>
+        <Route path="/posts/:id"      component={ShowPost}/>
       </Switch>
     );
   }
