+++ assets/js/App/Main/Footer/AuthPinapple/AuthPinapple.jsx
